added categoryController.js, modified userController.js
